reorder rows to get zeros of of diagonal entries.
In most cases, it should be easy to get zero entries out of the diagonal, but if the matrix has enough zeros, a full search may be necessary.
This is my implementation to do that, though I haven't run the code or tested it for syntax or correctness. 
this one fixes the dfs.